20181113-02 V1.8 （1）修复图片清理BUG

+++ src/main/java/com/xukai/netshop/scheduledTasks/FastdfsPicTask.java
@@ -3,10 +3,12 @@ package com.xukai.netshop.scheduledTasks;
 import com.xukai.netshop.config.BaseUrlConfig;
 import com.xukai.netshop.config.FastDfsConfig;
 import com.xukai.netshop.dataobject.ProductInfo;
+import com.xukai.netshop.enums.ResultEnum;
 import com.xukai.netshop.exception.SellException;
 import com.xukai.netshop.service.PictureService;
 import com.xukai.netshop.service.ProductService;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.Page;
@@ -42,12 +44,21 @@ public class FastdfsPicTask {
     /**
      * 每六小时执行一次旧图片清理
      */
-    @Scheduled(cron = "0 0 */6 * * ?")
+    // @Scheduled(cron = "0 0 */6 * * ?")
+    @Scheduled(cron = "10 */1 * * * ?")
     public void clearPicFromFastdfs() {
         log.info("【定时任务 - 清理过期图片】每六小时执行一次");
         String savedPicUrlsStr = "";
         BufferedReader reader = null;
         File file = new File(fastDfsConfig.getPsicSaveLogUrl());
+        if (!file.exists()) {
+            try {
+                file.createNewFile();
+            } catch (IOException e) {
+                log.error(ResultEnum.PICTURE_SAVED_LOG_CREATE_FAIL.getMessage());
+                e.printStackTrace();
+            }
+        }
         try {
             int fileLength = (int) file.length();
             reader = new BufferedReader(new FileReader(file.getPath()));
@@ -55,7 +66,7 @@ public class FastdfsPicTask {
             reader.read(c);
             savedPicUrlsStr = new String(c);
         } catch (Exception e) {
-            log.error("【卖家端清理过期图片】发生异常{}", e);
+            log.error("【卖家端清理过期图片】发生异常，{}", ResultEnum.PICTURE_SAVED_LOG_LOAD_FAIL.getMessage());
         } finally {
             try {
                 reader.close();
@@ -79,9 +90,11 @@ public class FastdfsPicTask {
         savedPicUrls.removeAll(existPicUrls);
         for (String toDelPicUrl : savedPicUrls) {
             try {
-                pictureService.deleteFile(toDelPicUrl);
+                if (StringUtils.isNotEmpty(toDelPicUrl)) {
+                    pictureService.deleteFile(toDelPicUrl);
+                }
             } catch (SellException e) {
-                log.error("【卖家端清理过期图片】发生异常{}", e);
+                log.error("【卖家端清理过期图片】发生异常，{}", ResultEnum.PICTURE_DELETE_FAIL.getMessage());
             }
         }
         // 将现存的PicUrl存入文本
@@ -94,12 +107,12 @@ public class FastdfsPicTask {
                 writer.write(existPicUrl + "\r\n");
             }
         } catch (IOException e) {
-            log.error("【卖家端清理过期图片】发生异常{}", e);
+            log.error("【卖家端清理过期图片】发生异常{}", ResultEnum.PICTURE_SAVED_LOG_REFRESH_FAIL.getMessage());
         } finally {
             try {
                 writer.close();
             } catch (IOException e) {
-                log.error("【卖家端清理过期图片】发生异常{}", e);
+                log.error("【卖家端清理过期图片】发生异常{}", ResultEnum.PICTURE_SAVED_LOG_REFRESH_FAIL.getMessage());
             }
         }
     }
